Add code editor layout
