+++ app/editor/page.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function CodeEditor() {
+  return <div>Code Editor</div>;
+}
