+++ components/logo/index.tsx
@@ -1,0 +1,9 @@
+import { NextLogo } from './next-logo';
+
+export function Logo() {
+  return (
+    <div className="border-white-30 group-hover:border-white-50 h-7 w-7 rounded-full border">
+      <NextLogo />
+    </div>
+  );
+}
