--- components/user-avatar.tsx
@@ -1,86 +1,0 @@
-import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar';
-import {
-  DropdownMenu,
-  DropdownMenuItem,
-  DropdownMenuContent,
-  DropdownMenuShortcut,
-  DropdownMenuGroup,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-  DropdownMenuSub,
-  DropdownMenuSubTrigger,
-  DropdownMenuSubContent,
-  DropdownMenuPortal,
-} from '@/components/ui/dropdown-menu';
-import { useTheme } from 'next-themes';
-
-interface UserAvatarProps {
-  name: string;
-  email: string;
-}
-
-export function UserAvatar({ name, email }: UserAvatarProps) {
-  const { setTheme } = useTheme();
-  return (
-    <div className="fixed bottom-0 pb-4">
-      <DropdownMenu>
-        <DropdownMenuTrigger>
-          <div className="flex w-full items-center text-left">
-            <Avatar className="h-8 w-8">
-              <AvatarImage src={`https://avatar.tobi.sh/${name}`} alt={name} />
-              <AvatarFallback className="text-xs">
-                {name && name.charAt(0).toUpperCase()}
-              </AvatarFallback>
-            </Avatar>
-            <div className="ml-2 text-left">
-              <p className="font-md text-sm">{name}</p>
-              <p className="text-xs text-gray-500">{email}</p>
-            </div>
-          </div>
-        </DropdownMenuTrigger>
-
-        <DropdownMenuContent className="ml-2 w-56">
-          <DropdownMenuLabel>My Account</DropdownMenuLabel>
-          <DropdownMenuSeparator />
-          <DropdownMenuGroup>
-            <DropdownMenuItem>
-              Profile
-              <DropdownMenuShortcut>⇧⌘P</DropdownMenuShortcut>
-            </DropdownMenuItem>
-            <DropdownMenuItem>
-              Billing
-              <DropdownMenuShortcut>⌘B</DropdownMenuShortcut>
-            </DropdownMenuItem>
-            <DropdownMenuItem>
-              Settings
-              <DropdownMenuShortcut>⌘S</DropdownMenuShortcut>
-            </DropdownMenuItem>
-          </DropdownMenuGroup>
-          <DropdownMenuSeparator />
-          <DropdownMenuSub>
-            <DropdownMenuSubTrigger>Theme</DropdownMenuSubTrigger>
-            <DropdownMenuPortal>
-              <DropdownMenuSubContent>
-                <DropdownMenuItem onClick={() => setTheme('light')}>
-                  Light
-                </DropdownMenuItem>
-                <DropdownMenuItem onClick={() => setTheme('dark')}>
-                  Dark
-                </DropdownMenuItem>
-                <DropdownMenuItem onClick={() => setTheme('system')}>
-                  System
-                </DropdownMenuItem>
-              </DropdownMenuSubContent>
-            </DropdownMenuPortal>
-          </DropdownMenuSub>
-          <DropdownMenuSeparator />
-          <DropdownMenuItem>
-            Log out
-            <DropdownMenuShortcut>⇧⌘Q</DropdownMenuShortcut>
-          </DropdownMenuItem>
-        </DropdownMenuContent>
-      </DropdownMenu>
-    </div>
-  );
-}
